requirements.txt
asgiref==3.4.1
Django==3.2.19
djangorestframework==3.14.0
djangorestframework-simplejwt==4.4.0
pkg-resources==0.0.0
PyJWT==2.4.0
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.1.1
